# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop Project API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/content/post": {
            "get": {
                "tags": [
                    "Admin/Content/Post"
                ],
                "summary": "List Posts",
                "description": "Retrieve a paginated list of posts.",
                "operationId": "7b6d7e64dd3017bd3bdcc884719523db",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin/Content/Post"
                ],
                "summary": "Create Post",
                "operationId": "9016101338d8907f72504ba41a0f000d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "summary",
                                    "body",
                                    "category_id",
                                    "image",
                                    "commentable",
                                    "published_at",
                                    "tags"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "هوش مصنوعی چگونه آینده توسعه نرم‌افزار را متحول می‌کند؟"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "هوش مصنوعی در حال تغییر روش توسعه، تست و نگهداری نرم‌افزارها است."
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "در سال‌های اخیر، هوش مصنوعی به یکی از مهم‌ترین فناوری‌های تأثیرگذار بر صنعت نرم‌افزار تبدیل شده است."
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "Post image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "commentable": {
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "published_at": {
                                        "description": "Unix timestamp",
                                        "type": "integer",
                                        "example": 1782331200
                                    },
                                    "tags": {
                                        "type": "string",
                                        "example": "هوش مصنوعی-فناوری-برنامه نویسی-AI-تکنولوژی"
                                    },
                                    "status": {
                                        "description": "Post status (optional)",
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostCategory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "تکنولوژی"
                    },
                    "description": {
                        "type": "string",
                        "example": "بررسی جدیدترین تکنولوژی‌ها"
                    },
                    "image": {
                        "type": "string",
                        "example": "images/category.png"
                    },
                    "tags": {
                        "type": "string",
                        "example": "tech"
                    }
                },
                "type": "object"
            },
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "هوش مصنوعی چگونه آینده توسعه نرم‌افزار را متحول می‌کند؟"
                    },
                    "summary": {
                        "type": "string",
                        "example": "هوش مصنوعی در حال تغییر روش توسعه، تست و نگهداری نرم‌افزارها است."
                    },
                    "body": {
                        "type": "string",
                        "example": "هوش مصنوعی در حال تغییر روش توسعه، تست و نگهداری نرم‌افزارها است."
                    },
                    "image": {
                        "type": "string",
                        "example": "images/post/2026/06/24/image.png"
                    },
                    "commentable": {
                        "type": "integer",
                        "example": 1
                    },
                    "published_at": {
                        "type": "string",
                        "example": "2026-06-24 23:00:00"
                    },
                    "author": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/PostCategory"
                    }
                },
                "type": "object"
            },
            "ForbiddenError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
...
            "UnauthenticatedError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": "?page=1",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "?page=3",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}